Move APCUPSd to separate file
