+++ homeassistant/components/apcupsd/const.py
@@ -1,0 +1,4 @@
+"""Constants for APCUPSd component."""
+from typing import Final
+
+DOMAIN: Final = "apcupsd"
